Split up golden tests and add test for multiple products
