+++ pkg/invoice/testdata/products.json
@@ -1,0 +1,144 @@
+[
+	{
+		"Tenant": {
+			"Source": "my-tenant",
+			"Target": ""
+		},
+		"PeriodStart": "2022-03-01T00:00:00Z",
+		"PeriodEnd": "2022-03-31T00:00:00Z",
+		"Categories": [
+			{
+				"Source": "my-cluster:my-namespace",
+				"Target": "",
+				"Items": [
+					{
+						"Description": "test description",
+						"Source": "my-product:my-cluster:my-tenant",
+						"Target": "",
+						"Quantity": 9072,
+						"QuantityMin": 42,
+						"QuantityAvg": 42,
+						"QuantityMax": 42,
+						"Unit": "tps",
+						"PricePerUnit": 3,
+						"Discount": 0,
+						"Total": 27216,
+						"SubItems": [
+							{
+								"Description": "A sub query of Test",
+								"Quantity": 864,
+								"QuantityMin": 4,
+								"QuantityAvg": 4,
+								"QuantityMax": 4,
+								"Unit": "tps"
+							}
+						]
+					}
+				],
+				"Total": 27216
+			},
+			{
+				"Source": "my-cluster:other-namespace",
+				"Target": "",
+				"Items": [
+					{
+						"Description": "test description",
+						"Source": "my-product:my-cluster:my-tenant",
+						"Target": "",
+						"Quantity": 9072,
+						"QuantityMin": 42,
+						"QuantityAvg": 42,
+						"QuantityMax": 42,
+						"Unit": "tps",
+						"PricePerUnit": 3,
+						"Discount": 0,
+						"Total": 27216,
+						"SubItems": [
+							{
+								"Description": "A sub query of Test",
+								"Quantity": 864,
+								"QuantityMin": 4,
+								"QuantityAvg": 4,
+								"QuantityMax": 4,
+								"Unit": "tps"
+							}
+						]
+					}
+				],
+				"Total": 27216
+			},
+			{
+				"Source": "other-cluster:my-namespace",
+				"Target": "",
+				"Items": [
+					{
+						"Description": "test description",
+						"Source": "my-product:*:my-tenant",
+						"Target": "",
+						"Quantity": 9072,
+						"QuantityMin": 42,
+						"QuantityAvg": 42,
+						"QuantityMax": 42,
+						"Unit": "tps",
+						"PricePerUnit": 2,
+						"Discount": 0,
+						"Total": 18144,
+						"SubItems": [
+							{
+								"Description": "A sub query of Test",
+								"Quantity": 864,
+								"QuantityMin": 4,
+								"QuantityAvg": 4,
+								"QuantityMax": 4,
+								"Unit": "tps"
+							}
+						]
+					}
+				],
+				"Total": 18144
+			}
+		],
+		"Total": 72576
+	},
+	{
+		"Tenant": {
+			"Source": "other-tenant",
+			"Target": ""
+		},
+		"PeriodStart": "2022-03-01T00:00:00Z",
+		"PeriodEnd": "2022-03-31T00:00:00Z",
+		"Categories": [
+			{
+				"Source": "my-cluster:my-namespace",
+				"Target": "",
+				"Items": [
+					{
+						"Description": "test description",
+						"Source": "my-product",
+						"Target": "",
+						"Quantity": 4968,
+						"QuantityMin": 23,
+						"QuantityAvg": 23,
+						"QuantityMax": 23,
+						"Unit": "tps",
+						"PricePerUnit": 1,
+						"Discount": 0,
+						"Total": 4968,
+						"SubItems": [
+							{
+								"Description": "A sub query of Test",
+								"Quantity": 432,
+								"QuantityMin": 2,
+								"QuantityAvg": 2,
+								"QuantityMax": 2,
+								"Unit": "tps"
+							}
+						]
+					}
+				],
+				"Total": 4968
+			}
+		],
+		"Total": 4968
+	}
+]
